Move message batching and publishing out of the Twitch proxy
